Get rid of `babel.util.FixedOffsetTimezone` in favor of `pytz.FixedOffset`.
Fixes #215

Fixes #217

Fixes #218
