+++ tests/test_util.py
@@ -11,8 +11,6 @@
 # individuals. For the exact contribution history, see the revision
 # history and logs, available at http://babel.edgewall.org/log/.
 
-import unittest
-
 from babel import util
 from babel._compat import BytesIO
 
@@ -42,18 +40,6 @@ def test_odict_pop():
         assert False
     except KeyError:
         assert True
-
-
-class FixedOffsetTimezoneTestCase(unittest.TestCase):
-
-    def test_zone_negative_offset(self):
-        self.assertEqual('Etc/GMT-60', util.FixedOffsetTimezone(-60).zone)
-
-    def test_zone_zero_offset(self):
-        self.assertEqual('Etc/GMT+0', util.FixedOffsetTimezone(0).zone)
-
-    def test_zone_positive_offset(self):
-        self.assertEqual('Etc/GMT+330', util.FixedOffsetTimezone(330).zone)
 
 
 parse_encoding = lambda s: util.parse_encoding(BytesIO(s.encode('utf-8')))
